fix issue

+++ source/admin-service/src/main/java/org/vision/service/admin/service/vo/VisionClientVO.java
@@ -15,7 +15,7 @@ public interface VisionClientVO {
 
     String getName();
 
-    Byte getGender();
+    Integer getGender();
 
     Integer getAge();
